Ajout de typescript dans le projet et adaptations
