+++ package.json
@@ -10,22 +10,26 @@
   },
   "dependencies": {
     "ioredis": "^5.3.2",
-    "next": "^14.1.3",
+    "next": "14.2.1",
     "next-auth": "^4.24.7",
-    "react": "^18.2.0",
+    "react": "^18",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18",
     "ssh2": "^1.15.0",
     "swr": "^2.2.5"
   },
   "devDependencies": {
+    "typescript": "^5",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "autoprefixer": "^10",
     "eslint": "^8",
-    "eslint-config-next": "13.5.4",
+    "eslint-config-next": "14.2.1",
     "nextjs-node-loader": "^1.1.5",
     "postcss": "^8",
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.13",
-    "tailwindcss": "^3"
+    "tailwindcss": "^3.4.1"
   }
 }
