+++ src/app/layout.tsx
@@ -1,16 +1,22 @@
-import "./globals.css";
+import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { Providers } from "./Providers";
+import "./globals.css";
+import React from "react";
+import { Providers } from "@/app/Providers";
 import SessionGuard from "@/components/SessionGuard";
 
 const inter = Inter({ subsets: ["latin"] });
 
-export const metadata = {
+export const metadata: Metadata = {
   title: "Dashboard-listing",
   description: "Liste des liens dashboard",
 };
 
-export default function RootLayout({ children }) {
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   return (
     <html lang="en">
       <body className={inter.className}>
